# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Telegram Bot Dashboard API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface GitHubConnectBody {
  /** GitHub repository HTTPS URL */
  repoUrl: string;
  /** GitHub Personal Access Token */
  token: string;
  /** Branch to sync (defaults to main) */
  branch?: string;
}

export interface GitHubConnectResult {
  success: boolean;
  message: string;
  repoName?: string;
  branch?: string;
}

export interface GitHubStatus {
  connected: boolean;
  repoUrl?: string;
  repoName?: string;
  branch?: string;
  lastSync?: string | null;
  lastPush?: string | null;
  localPath?: string | null;
}

export interface SyncResult {
  success: boolean;
  message: string;
  filesChanged?: string[];
  timestamp?: string;
}

export interface PushBody {
  commitMessage: string;
}

export interface BotFileMeta {
  path: string;
  name: string;
  size: number;
  lastModified: string;
}

export interface BotFileList {
  files: BotFileMeta[];
}

export interface BotFileContent {
  path: string;
  content: string;
}

export interface BotConfig {
  hasTelegramToken: boolean;
  hasMongoUri: boolean;
  hasGitHubToken: boolean;
  repoUrl?: string | null;
}

export interface BotConfigBody {
  telegramToken?: string | null;
  mongoUri?: string | null;
}

export interface BotLogs {
  lines: string[];
  timestamp: string;
}

export type GetBotFileContentParams = {
  path: string;
};