# lib/api-zod/src/generated/types/botConfig.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Telegram Bot Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface BotConfig {
  hasTelegramToken: boolean;
  hasMongoUri: boolean;
  hasGitHubToken: boolean;
  repoUrl?: string | null;
}