# lib/api-zod/src/generated/types/botConfigBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Telegram Bot Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface BotConfigBody {
  telegramToken?: string | null;
  mongoUri?: string | null;
}