# lib/api-zod/src/generated/types/botLogs.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Telegram Bot Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface BotLogs {
  lines: string[];
  timestamp: string;
}