# lib/api-zod/src/generated/types/gitHubConnectBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Telegram Bot Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface GitHubConnectBody {
  /** GitHub repository HTTPS URL */
  repoUrl: string;
  /** GitHub Personal Access Token */
  token: string;
  /** Branch to sync (defaults to main) */
  branch?: string;
}