# lib/api-zod/src/generated/types/gitHubStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Telegram Bot Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface GitHubStatus {
  connected: boolean;
  repoUrl?: string;
  repoName?: string;
  branch?: string;
  lastSync?: string | null;
  lastPush?: string | null;
  localPath?: string | null;
}